--- views/scripts/index-controller.js
@@ -1,9 +1,0 @@
-
-
-
-export const navigationLeftToHomePage = () => {
-    document.querySelector(".navigation__left").addEventListener("click", function() {
-    const homePage = `http://${window.location.host}`;
-    window.location.href = homePage;
-    }
-)};
