--- views/scripts/index.js
@@ -1,11 +1,0 @@
-import { fetchAttractionsData } from "./index-attractions.js";
-import { listBarApi } from "./index-listbar.js";
-import { searchInputApi } from "./index-search.js";
-
-
-
-window.addEventListener("DOMContentLoaded", function(){
-    fetchAttractionsData();
-    listBarApi();
-    searchInputApi();
-});
